Write a better texture loader
