+++ src/loaders/textures.js
@@ -1,44 +1,28 @@
-import planets from '../constants/planets/_types'
-import resources from '../constants/resources/_types'
+import { Texture } from 'pixi.js'
 
-const texturePrefix = 'assets/img'
+const prefix = 'assets/img'
 
-const textures = []
-
-planets.forEach(element => {
-  const texture = {
-    name: `Planets_${element}`,
-    url: `${texturePrefix}/Planets/${element}.pngs`
+const planetTypes = [
+  {
+    name: 'sun',
+    url: 'stars',
+    num: 1,
+    type: 'png'
+  },
+  {
+    name: 'Habitital_Planet',
+    url: 'planets',
+    num: 1,
+    type: 'png'
   }
-  textures.push(texture)
-})
+]
 
-resources.forEach(element => {
-  const texture = {
-    name: `Resources_${element}`,
-    url: `${texturePrefix}/Resources/${element}.png`
+const textures = {}
+
+planetTypes.forEach(planetType => {
+  for (let i = 1; i !== planetType.num; i++) {
+    textures[planetType.name + i] = Texture.from(`${prefix}/${planetType.url}/${planetType.name}${i}.${planetType.prefix}`)
   }
-  textures.push(texture)
-})
-
-/* stars.forEach(element => {
-  const texture = {
-    name: `Stars_${element}`,
-    url: `${texturePrefix}/Stars/${element}.png`
-  }
-  textures.push(texture)
-}) */
-
-//! Probibly could be improved somehow
-textures.push({
-  name: 'Atmosphere',
-  url: `${texturePrefix}/Atmosphere1.svg`
-})
-
-//! Temprary code
-textures.push({
-  name: 'Stars_Sun',
-  url: `${texturePrefix}/stars/Sun1.png`
 })
 
 export default textures
